added more validation before CRUD operations in back- and frontend

+++ swtcamper/src/main/java/swtcamper/backend/services/BookingService.java
@@ -40,22 +40,39 @@ public class BookingService {
   public Booking update(
     Long bookingID,
     LocalDate startDate,
-    LocalDate endDate,
-    boolean active
+    LocalDate endDate
   ) throws GenericServiceException {
     // Search for booking in database
     Optional<Booking> bookingOptional = bookingRepository.findById(bookingID);
     if (bookingOptional.isPresent()) {
       // Booking found so update can be made
       Booking booking = bookingOptional.get();
+      if (booking.isActive()) {
+        throw new GenericServiceException("The booking with ID " + bookingID + " cannot be deleted since it is still active.");
+      }
       booking.setStartDate(startDate);
       booking.setEndDate(endDate);
-      booking.setActive(active);
       // Save update back to database
       return bookingRepository.save(booking);
     }
     throw new GenericServiceException(
       "Booking not found. Update not possible."
+    );
+  }
+
+  public Booking activate(Long bookingID) throws GenericServiceException {
+    // Search for booking in database
+    Optional<Booking> bookingOptional = bookingRepository.findById(bookingID);
+    if (bookingOptional.isPresent()) {
+      // Booking found so update can be made
+      Booking booking = bookingOptional.get();
+      // Update by setting active = true
+      booking.setActive(true);
+      // Save update back to database
+      return bookingRepository.save(booking);
+    }
+    throw new GenericServiceException(
+            "Booking not found. Activation not possible."
     );
   }
 
@@ -66,13 +83,9 @@ public class BookingService {
       // Booking found so update can be made
       Booking booking = bookingOptional.get();
       // Update by setting active = false
+      booking.setActive(false);
       // Save update back to database
-      return this.update(
-          booking.getId(),
-          booking.getStartDate(),
-          booking.getEndDate(),
-          false
-        );
+      return bookingRepository.save(booking);
     }
     throw new GenericServiceException(
       "Booking not found. Deactivation not possible."
@@ -80,10 +93,19 @@ public class BookingService {
   }
 
   public void delete(Long bookingID) throws GenericServiceException {
-    try {
-      bookingRepository.deleteById(bookingID);
-    } catch (IllegalArgumentException e) {
-      throw new GenericServiceException("The passed ID is not available: " + e);
+    Optional<Booking> bookingOptional = bookingRepository.findById(bookingID);
+    if (bookingOptional.isPresent()) {
+      // Booking found
+      Booking booking = bookingOptional.get();
+      // can't delete if booking is active
+      if (booking.isActive()) {
+        throw new GenericServiceException("The booking with ID " + bookingID + " cannot be deleted since it is still active.");
+      }
+      try {
+        bookingRepository.deleteById(bookingID);
+      } catch (IllegalArgumentException e) {
+        throw new GenericServiceException("The passed ID is not available: " + e);
+      }
     }
   }
